requirements.txt
together>=1.4.1
python-dotenv>=1.0.0
google-generativeai>=0.3.0
requests>=2.31.0
Pillow>=11.1.0
streamlit>=1.31.0
pytest>=7.4.0 